verl-tool_060/requirements.txt
accelerate==1.6.0
acecoder @ git+https://github.com/TIGER-AI-Lab/AceCoder.git@ad017bcef56a95e5b06dc8da564c8a8ab6e3d62e
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiohttp-cors==0.8.1
aiosignal==1.3.2
airportsdata==20250224
annotated-types==0.7.0
anthropic==0.50.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
appdirs==1.4.4
astor==0.8.1
async-timeout==5.0.1
attrs==25.3.0
blake3==1.0.4
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
codetiming==1.4.0
colorful==0.5.6
compressed-tensors==0.9.3
cupy-cuda12x==13.4.1
datasets==3.5.0
deprecated==1.2.18
depyf==0.18.0
dill==0.4.0
diskcache==5.6.3
distlib==0.3.9
distro==1.9.0
dnspython==2.7.0
docker-pycreds==0.4.0
einops==0.8.1
email-validator==2.2.0
evalplus==0.3.1
exceptiongroup==1.2.2
fastapi==0.115.12
fastapi-cli==0.0.7
fastrlock==0.8.3
filelock==3.18.0
fire==0.7.0
flash-attn==2.7.4.post1
frozenlist==1.6.0
fsspec==2025.3.2
gguf==0.16.2
gitdb==4.0.12
gitpython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.168.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
hf-xet==1.0.5
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.30.2
hydra-core==1.3.2
idna==3.10
importlib-metadata==8.0.0
interegular==0.3.3
jinja2==3.1.6
jiter==0.9.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
lark==1.2.2
latex2sympy2-extended==1.10.1
llguidance==0.7.19
llvmlite==0.44.0
lm-format-enforcer==0.10.11
markdown-it-py==3.0.0
markupsafe==3.0.2
math-verify==0.7.0
mdurl==0.1.2
mistral-common==1.5.4
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.19.0
multidict==6.4.3
multipledispatch==1.0.0
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.4
numba==0.61.2
numpy==2.2.5
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-ml-py==12.570.86
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
nvitop==1.5.0
omegaconf==2.3.0
openai==1.76.0
opencensus==0.11.4
opencensus-context==0.1.3
opencv-python-headless==4.11.0.86
opentelemetry-api==1.26.0
opentelemetry-exporter-otlp==1.26.0
opentelemetry-exporter-otlp-proto-common==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
opentelemetry-exporter-otlp-proto-http==1.26.0
opentelemetry-proto==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
opentelemetry-semantic-conventions-ai==0.4.3
orjson==3.10.16
outlines==0.1.11
outlines-core==0.1.26
packaging==25.0
pandas==2.2.3
partial-json-parser==0.2.1.1.post5
peft==0.15.2
pillow==11.2.1
platformdirs==4.3.7
prometheus-client==0.21.1
prometheus-fastapi-instrumentator==7.1.0
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
psutils==3.3.9
puremagic==1.28
py-cpuinfo==9.0.0
py-spy==0.4.0
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pybind11==2.13.6
pycountry==24.6.1
pydantic==2.11.3
pydantic-core==2.33.1
pygments==2.19.1
pylatexenc==2.10
pyparsing==3.2.3
pypdf==5.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==3.3.0
python-multipart==0.0.20
pytz==2025.2
pyyaml==6.0.2
pyzmq==26.4.0
ray==2.43.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rich-toolkit==0.14.3
rpds-py==0.24.0
rsa==4.9.1
safetensors==0.5.3
scipy==1.15.2
sentencepiece==0.2.0
sentry-sdk==2.27.0
setproctitle==1.3.5
setuptools==80.0.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.46.2
stop-sequencer==1.2.3
sympy==1.13.1
tempdir==0.7.1
tensordict==0.6.2
termcolor==3.0.1
tiktoken==0.9.0
timeout-decorator==0.5.0
tokenizers==0.21.1
torch==2.6.0
torchaudio==2.6.0
torchdata==0.11.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.51.3
tree-sitter==0.24.0
tree-sitter-python==0.23.6
triton==3.2.0
typer==0.15.2
typing-extensions==4.13.2
typing-inspection==0.4.0
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
-e file:///home/aiops/jiangdf/Workspace/verl-tool/verl
-e file:///home/aiops/jiangdf/Workspace/verl-tool
virtualenv==20.30.0
vllm==0.8.4
wandb==0.19.10
watchfiles==1.0.5
websockets==15.0.1
wget==3.2
wrapt==1.17.2
xformers==0.0.29.post2
xgrammar==0.1.18
xxhash==3.5.0
yarl==1.20.0
zipp==3.21.0
